requirements/production.txt
-r base.txt
dj-database-url==0.3.0
psycopg2==2.6
newrelic==2.50.0.39
django-celery==3.1.16
gunicorn==19.3.0
django-redis-cache==0.13.1
redis==2.10.3
